Added control strip, removed selector bar and HTML pane, and added support for template data fetching from the back end.

+++ cmd/winfe/cxicon.cpp
@@ -128,9 +128,12 @@ void CRDFImage::RemoveListenerForSpecificResource(CCustomImageObject *pObject, H
 
 void CRDFImage::DestroyContext()
 {
-	iconContext->DeleteContextDC();
-	iconContext->NiceDestruction();
-	iconContext = NULL;
+	if (iconContext)
+	{
+		iconContext->DeleteContextDC();
+		iconContext->NiceDestruction();
+		iconContext = NULL;
+	}
 
 	resourceList.RemoveAll();
 }
@@ -172,12 +175,14 @@ void Icon_GetUrlExitRoutine(URL_Struct *pUrl, int iStatus, MWContext *pContext)
 		}
 	}
 	// standard free-a-URL-with-Windows-DDE-gunk mechanism:
+	
 	if (NCAPIDATA(pUrl))
 		NCAPIDATA(pUrl)->EndProgress();
 	if (!NCAPIDATA(pUrl) || NCAPIDATA(pUrl)->CanFreeUrl()) {
 		FEU_DeleteUrlData(pUrl, NULL);
 		NET_FreeURLStruct(pUrl);
 	}
+	
 }
 
 static BOOL IsImageMimeType(const CString& theFormat)
@@ -270,6 +275,7 @@ void CRDFImage::ProcessIcon()
 void CRDFImage::CompleteCallback()
 {
 	m_bCompletelyLoaded = TRUE;
+	m_bFrameLoaded = TRUE;
 	while (!resourceList.IsEmpty())
 	{
 		CIconCallbackInfo* callback = (CIconCallbackInfo*)(resourceList.RemoveHead());
